requirements.txt
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
et-xmlfile==1.1.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gspread==6.0.0
idna==3.6
numpy==1.26.3
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.2.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
StrEnum==0.4.15
tzdata==2023.4
urllib3==2.2.0
xlrd==2.0.1
